requirements.txt
python-telegram-bot[webhooks]>=21.0
sqlalchemy>=2.0.36
python-dotenv>=1.0.1
httpx>=0.28.1
aiohttp>=3.9.1
psycopg>=3.1.18
psycopg2-binary>=2.9.9 